Use double-precision to store split thresholds

+++ include/treelite/c_api.h
@@ -338,7 +338,7 @@ TREELITE_DLL int TreeliteTreeBuilderSetNumericalTestNode(
                                              TreeBuilderHandle handle,
                                              int node_key, unsigned feature_id,
                                              const char* opname,
-                                             float threshold, int default_left,
+                                             double threshold, int default_left,
                                              int left_child_key,
                                              int right_child_key);
 /*!
@@ -375,7 +375,7 @@ TREELITE_DLL int TreeliteTreeBuilderSetCategoricalTestNode(
  */
 TREELITE_DLL int TreeliteTreeBuilderSetLeafNode(TreeBuilderHandle handle,
                                                 int node_key,
-                                                float leaf_value);
+                                                double leaf_value);
 /*!
  * \brief Turn an empty node into a leaf vector node
  * The leaf vector (collection of multiple leaf weights per leaf node) is
@@ -389,7 +389,7 @@ TREELITE_DLL int TreeliteTreeBuilderSetLeafNode(TreeBuilderHandle handle,
  */
 TREELITE_DLL int TreeliteTreeBuilderSetLeafVectorNode(TreeBuilderHandle handle,
                                                       int node_key,
-                                                      const float* leaf_vector,
+                                                      const double* leaf_vector,
                                                       size_t leaf_vector_len);
 /*!
  * \brief Create a new model builder
