+++ src/c_api/c_api.cc
@@ -389,7 +389,7 @@ int TreeliteTreeBuilderSetRootNode(TreeBuilderHandle handle, int node_key) {
 int TreeliteTreeBuilderSetNumericalTestNode(TreeBuilderHandle handle,
                                             int node_key, unsigned feature_id,
                                             const char* opname,
-                                            float threshold, int default_left,
+                                            double threshold, int default_left,
                                             int left_child_key,
                                             int right_child_key) {
   API_BEGIN();
@@ -430,7 +430,7 @@ int TreeliteTreeBuilderSetCategoricalTestNode(
 }
 
 int TreeliteTreeBuilderSetLeafNode(TreeBuilderHandle handle, int node_key,
-                                   float leaf_value) {
+                                   double leaf_value) {
   API_BEGIN();
   auto builder = static_cast<frontend::TreeBuilder*>(handle);
   CHECK(builder) << "Detected dangling reference to deleted TreeBuilder object";
@@ -441,7 +441,7 @@ int TreeliteTreeBuilderSetLeafNode(TreeBuilderHandle handle, int node_key,
 
 int TreeliteTreeBuilderSetLeafVectorNode(TreeBuilderHandle handle,
                                          int node_key,
-                                         const float* leaf_vector,
+                                         const double* leaf_vector,
                                          size_t leaf_vector_len) {
   API_BEGIN();
   auto builder = static_cast<frontend::TreeBuilder*>(handle);
